Separated out microscope specific SEM control classes
